move plugin code to entry file
